% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_framework}
\alias{api_framework}
\title{Find out the API framework}
\usage{
api_framework(req, res)
}
\arguments{
\item{req}{The request object.}

\item{res}{The response object.}
}
\value{
A character vector of length 1.
Possible values: \code{"plumber"}, \code{"RestRserve"} (R package names);
\code{"plumber"} is assumed when the API framework cannot be determined.
}
\description{
Find out the API framework
}
