% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_status_codes.R
\docType{data}
\name{http_status_codes}
\alias{http_status_codes}
\title{Data frame with possible status codes and default messages}
\format{
An object of class \code{data.frame} with 62 rows and 3 columns.
}
\usage{
http_status_codes
}
\description{
Data frame with possible status codes and default messages
}
\keyword{datasets}
