% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_taxa.R
\name{is_rt}
\alias{is_rt}
\title{Categories of microbial taxa}
\usage{
is_rt(x, t_r = 0.001)
}
\arguments{
\item{x}{The $n$ by $m$ matrix of microbial abundance.}
}
\value{
A vector of microbial taxa belonging to each category.
}
\description{
This function classify microbial taxa into six categories
RT: rare taxa
MT: moderate taxa
AT: abundant taxa
CRT: conditionally rare taxa
CAT: conditionally abundant taxa
CRAT: conditionally rare or abundant taxa
}
\examples{
test <- runif(10,0,1)
print(test)


}
