% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Message to STDOUT or STDERR}
\usage{
msg(title = "", message = "", level = "INFO", format = NULL, digits = NULL)
}
\arguments{
\item{title}{Title.}

\item{message}{Message.}

\item{level}{Log level.}

\item{format}{Log format: \code{"PLAIN"} (default), \code{"JSON"}, \code{"CSV"}..}

\item{digits}{Digits for seconds (default \code{3L}).}
}
\value{
\code{TRUE} or \code{FALSE}: did a log event happened?
A side effect is a log message to STDOUT or STDERR.
}
\description{
Message to STDOUT or STDERR
}
\examples{
msg("Success", "We did it!")
msg("Success", "We did it!", "SUCCESS")
msg("Crap", "Oh no ...", "ERROR")

msg("Success", "We did it!", "SUCCESS", format = "JSON")
msg("Success", "We did it!", format = "JSON")
msg("Crap", "Oh no ...", "ERROR", format = "JSON")

msg("Success", "We did it!", digits = 0)
msg("Success", "We did it!", digits = 6)
 
}
