% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{oneline}
\alias{oneline}
\title{Remove newlines and leading/trailing white space from a string}
\usage{
oneline(x)
}
\arguments{
\item{x}{A string, possibly a vector}
}
\value{
An atomic character vector.
}
\description{
Remove newlines and leading/trailing white space from a string
}
