% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_error.R
\name{http-messages}
\alias{http-messages}
\alias{http_error}
\alias{http_success}
\title{HTTP messages}
\usage{
http_error(status = 500L, message = NULL)

http_success(status = 200L, message = NULL, body = NULL)
}
\arguments{
\item{status}{HTTP status code.}

\item{message}{Error message.}

\item{body}{A list, additional values to be returned.}
}
\value{
\code{http_error} returns an error with a custom condition attribute.
\code{http_success} returns a list.
}
\description{
HTTP messages
}
\examples{
try(http_error())
try(http_error(400))
try(http_error(400, "Sorry"))

http_success()
http_success(201)
http_success(201, "Awesome")
http_success(201, "Awesome", list(name = "Jane", count = 6))


}
